Move requirements related code to requires.py
Restructure a bit for simplification.
