+++ test/test_py2pack.py
@@ -16,7 +16,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import pkg_resources
 import unittest
 from ddt import ddt, data, unpack
 
@@ -56,37 +55,6 @@ class Py2packTestCase(unittest.TestCase):
         d = {'license': value}
         py2pack._normalize_license(d)
         self.assertEqual(d['license'], expected_result)
-
-    @data(
-        ("pywin32>=1.0;sys_platform=='win32'  # PSF", False),
-        ("foobar", True),
-        ("foobar;python_version=='2.7'", True),
-        ("foobar;python_version=='3.5'", False),
-    )
-    @unpack
-    def test__requirement_filter_by_marker(self, req, expected):
-        pkg = pkg_resources.Requirement.parse(req)
-        self.assertEqual(py2pack._requirement_filter_by_marker(pkg), expected)
-
-    @data(
-        ("foobar>=1.0", ["foobar", ">=", "1.0"]),
-        ("foobar>=1.0,>2", ["foobar", ">=", "1.0"]),
-        ("foobar>=2,>1.0,<=3", ["foobar", ">", "1.0"]),
-        ("foobar>=2,>1.0,!=0.5", ["foobar", ">", "1.0"]),
-        ("foobar!=0.5", ["foobar"]),
-    )
-    @unpack
-    def test__requirement_find_lowest_possible(self, req, expected):
-        pkg = pkg_resources.Requirement.parse(req)
-        self.assertEqual(list(py2pack._requirement_find_lowest_possible(pkg)), expected)
-
-    @data(
-        (['six', 'monotonic>=0.1'], ['six', 'monotonic >= 0.1']),
-        (['monotonic>=1.0,>0.1'], ['monotonic > 0.1']),
-    )
-    @unpack
-    def test__requirements_sanitize(self, req_list, expected):
-        self.assertEqual(py2pack._requirements_sanitize(req_list), expected)
 
     @data(
         (
