+++ test/test_requires.py
@@ -16,14 +16,17 @@
 # limitations under the License.
 
 import os
+import pkg_resources
 import shutil
 import tempfile
 import unittest
+from ddt import ddt, data, unpack
 
 import py2pack
 import py2pack.utils
 
 
+@ddt
 class Py2packRequiresTestCase(unittest.TestCase):
 
     def setUp(self):
@@ -62,3 +65,34 @@ setuptools.setup(
         data = py2pack.requires._requires_from_setup_py_run(self.tmpdir)
         self.assertEqual(data, {'install_requires': ['foo', 'bar'],
                                 'tests_require': ['test']})
+
+    @data(
+        ("pywin32>=1.0;sys_platform=='win32'  # PSF", False),
+        ("foobar", True),
+        ("foobar;python_version=='2.7'", True),
+        ("foobar;python_version=='3.5'", False),
+    )
+    @unpack
+    def test__requirement_filter_by_marker(self, req, expected):
+        pkg = pkg_resources.Requirement.parse(req)
+        self.assertEqual(py2pack.requires._requirement_filter_by_marker(pkg), expected)
+
+    @data(
+        ("foobar>=1.0", ["foobar", ">=", "1.0"]),
+        ("foobar>=1.0,>2", ["foobar", ">=", "1.0"]),
+        ("foobar>=2,>1.0,<=3", ["foobar", ">", "1.0"]),
+        ("foobar>=2,>1.0,!=0.5", ["foobar", ">", "1.0"]),
+        ("foobar!=0.5", ["foobar"]),
+    )
+    @unpack
+    def test__requirement_find_lowest_possible(self, req, expected):
+        pkg = pkg_resources.Requirement.parse(req)
+        self.assertEqual(list(py2pack.requires._requirement_find_lowest_possible(pkg)), expected)
+
+    @data(
+        (['six', 'monotonic>=0.1'], ['six', 'monotonic >= 0.1']),
+        (['monotonic>=1.0,>0.1'], ['monotonic > 0.1']),
+    )
+    @unpack
+    def test__requirements_sanitize(self, req_list, expected):
+        self.assertEqual(py2pack.requires._requirements_sanitize(req_list), expected)
